requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn
scikit-learn
requests
joblib
